Fix remaining compilation errors after kotlinx-datetime upgrade
- FormatUtils.kt: add missing import for Month.number extension property
- TimeGrouping.kt: use kotlin.time.Clock for Clock.System access (type aliases don't forward nested objects)
- TimeGrouping.kt: convert toEpochDays() Long result to Int for comparison operators

+++ composeApp/src/commonMain/kotlin/com/klemfner/whoscalling/ui/common/utils/TimeGrouping.kt
@@ -1,6 +1,6 @@
 package com.klemfner.whoscalling.ui.common.utils
 
-import kotlinx.datetime.Clock
+import kotlin.time.Clock
 import kotlinx.datetime.Instant
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
@@ -14,7 +14,7 @@ fun getTimePeriod(timestampMillis: Long): TimePeriod {
     val now = Clock.System.now().toLocalDateTime(tz).date
     val date = Instant.fromEpochMilliseconds(timestampMillis).toLocalDateTime(tz).date
 
-    val daysDiff = now.toEpochDays() - date.toEpochDays()
+    val daysDiff = (now.toEpochDays() - date.toEpochDays()).toInt()
 
     return when {
         daysDiff == 0 -> TimePeriod.TODAY
